Split Code Run linters for errors
