+++ src/routes/FAQ.js
@@ -1,23 +1,7 @@
 import React from 'react';
-import PropTypes from 'prop-types'; // Import PropTypes for prop validation
 import './styles/faq.css';
 import OurPartners from '../components/ourPartners';
-
-// Define prop types for FAQItem
-function FAQItem({ question, answer }) {
-  return (
-    <details className="faq-item">
-      <summary className="faq-question">{question}</summary>
-      <p className="faq-answer">{answer}</p>
-    </details>
-  );
-}
-
-// Add PropTypes validation
-FAQItem.propTypes = {
-  question: PropTypes.string.isRequired,
-  answer: PropTypes.string.isRequired,
-};
+import FAQItem from '../components/logic/FAQItem';
 
 function FAQ() {
   return (
